Started to make the OS/X app
Struggling: starting the OS/X first no longer works.
Symptoms: there is NO LOGGING I NFORMATION when it's started
first, yet when it's started second there /is/ logging information.
Could it be that the OS/X app has a problem listening on the Red.Bus?
The same problem seems to retro-apply to Red when newly
constructed. Maybe a library problem?