Struggling to make the OS/X app  function properly
Idea: make AppleRed.app do no more than forward files to the ordinary non-app program that it will start in the background if necessary. The point is to avoid
the app itself trying to BE the server.

I don't quite understand what happened, since MacRed now seems to do the same thing as the aborted AppleDred.