(A), (F), (R) buttons now clear their attached text fields
Still irritated by focus-pushing glitch; and investigating.
CLUE: ctrl-tab moves the focus to the next of the four text fields. Well, not quite,
for on the minitexts, one must type a TAB before the focus actually swaps from the text it's leaving to that it's going to. Except when the target is the main text.